Add result() and result_ptr_mut()
When dealing with ffi functions that return 0 for success and -1 for error,
or NULL for error and <ptr> for success, these make it easier to convert
these into Rust Results.

This commit also checks the retval of blkid_new_probe_from_filename using
result_ptr_mut() as an example of usage.

Signed-off-by: Andy Grover <agrover@redhat.com>