Initial support for Cache, Dev, and Tag functions
I actually started writing a blkid wrapper lib from scratch. Luckily I
didn't start in the same place as yours, so we can merge stuff!

This is a rough first implementation to wrap blkid's cache, dev, and tag
APIs, which are all very similar in terms of being iteration-oriented and
mapping well to Iterators.

Signed-off-by: Andy Grover <agrover@redhat.com>