Fix compilation for 32-bit architectures
The topology functions return unsigned long, so we can keep the API
simple by just converting 32-bit values to u64.

For partitions_get_name(), which takes a size_t, converting a u64 on
32-bit architectures could fail, so it's better to adjust the API to
try to ensure overly large values can't be passed.

Technically, usize is uintptr_t, not size_t, and uintptr_t can be
bigger than size_t, so it's possible that the architecture could allow
overly big values to be passed to partitions_get_name(), but I'm not
aware of any Linux architectures where that's the case, and even if it
was, it would just result in a panic if too big a value was passed.
It's worth it to expose usize in the high-level API, rather than
libc::size_t.